simplifying enums

+++ SerenityLib/include/serenity/events/EventDispatcher.h
@@ -15,9 +15,9 @@ namespace Engine::priv {
     class EventDispatcher final {
         using ObserverVector = std::vector<Observer*>;
         private:
-            mutable std::mutex                                       m_Mutex;
-            std::array<ObserverVector, (size_t)EventType::_TOTAL>    m_Observers;
-            std::vector<std::pair<Observer*, size_t>>                m_UnregisteredObservers;
+            mutable std::mutex                                m_Mutex;
+            std::array<ObserverVector, EventType::_TOTAL>     m_Observers;
+            std::vector<std::pair<Observer*, size_t>>         m_UnregisteredObservers;
 
             [[nodiscard]] bool internal_has_duplicate(const Observer&, const ObserverVector&) const noexcept;
             void internal_dispatch_event(const Event&);
